Refactor email redux logic
