+++ .eslintrc.js
@@ -1,4 +1,8 @@
 module.exports = {
   root: true,
-  extends: '@react-native-community'
-}
+  extends: '@react-native-community',
+  rules: {
+    'react-hooks/exhaustive-deps': 'warn',
+    '@typescript-eslint/no-unused-vars': 'warn',
+  },
+};
