+++ src/store/adapters/emails.ts
@@ -1,0 +1,10 @@
+import { createEntityAdapter } from '@reduxjs/toolkit';
+import { Email } from '../types';
+
+export const emailAdapter = createEntityAdapter<Email>({
+  selectId: (email: Email) => email.emailId,
+  sortComparer: (a, b) =>
+    new Date(b.createdAt).valueOf() - new Date(a.createdAt).valueOf(),
+});
+
+export const emailSelectors = emailAdapter.getSelectors();
