+++ src/store/adapters/folders.ts
@@ -1,0 +1,8 @@
+import { createEntityAdapter } from '@reduxjs/toolkit';
+import { Folder } from '../types';
+
+export const folderAdapter = createEntityAdapter<Folder>({
+  selectId: folder => folder.folderId,
+});
+
+export const folderSelector = folderAdapter.getSelectors();
