+++ src/store/selectors/namespaces.ts
@@ -1,10 +1,10 @@
 import { RootState } from '../../store';
-import { createSelector } from '@reduxjs/toolkit';
 import { namespaceSelectors } from '../adapters/namespaces';
+import { createDeepEqualSelector } from './utils';
 
 export const namespaceSelector = (state: RootState) => state.namespaces;
 
-export const namespaceNameSelector = createSelector(
+export const namespaceNameSelector = createDeepEqualSelector(
   (state: RootState) => namespaceSelectors.selectAll(state.namespaces),
   namespaces => namespaces.map(namespace => namespace.name),
 );
