+++ src/store/selectors/utils.ts
@@ -1,0 +1,7 @@
+import { createSelectorCreator, defaultMemoize } from 'reselect';
+import isEqual from 'lodash/isEqual';
+
+export const createDeepEqualSelector = createSelectorCreator(
+  defaultMemoize,
+  isEqual,
+);
