+++ src/store/types/enums/Folders.ts
@@ -1,0 +1,7 @@
+export enum FoldersId {
+  inbox = 1,
+  drafts = 2,
+  sent = 3,
+  trash = 4,
+  aliases = 5,
+}
